// frontend/src/app/core/api/model/user-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponse { 
    id?: number;
    username?: string;
    role?: UserResponse.RoleEnum;
    status?: UserResponse.StatusEnum;
    forcePasswordChange?: boolean;
    personId?: number;
    firstName?: string;
    lastName?: string;
}
export namespace UserResponse {
    export const RoleEnum = {
        RoleAdmin: 'ROLE_ADMIN',
        RoleTeacher: 'ROLE_TEACHER',
        RoleStudent: 'ROLE_STUDENT',
        RoleParent: 'ROLE_PARENT'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
    export const StatusEnum = {
        Pending: 'PENDING',
        Active: 'ACTIVE',
        Suspended: 'SUSPENDED',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}