// frontend/src/app/core/api/model/admin-change-password-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminChangePasswordRequest { 
    newPassword: string;
    confirmNewPassword: string;
}